chore: restructure package hierarchy
