+++ core/src/main/java/io/mohamed/resolver/core/merge/AARMerger.java
@@ -37,7 +37,7 @@ import com.android.manifmerger.ManifestMerger2.MergeType;
 import com.android.manifmerger.MergingReport;
 import com.android.manifmerger.XmlDocument;
 import com.android.utils.ILogger;
-import io.mohamed.resolver.core.Util;
+import io.mohamed.resolver.core.util.Util;
 import io.mohamed.resolver.core.callback.DependencyResolverCallback;
 import io.mohamed.resolver.core.callback.DependencyResolverCallback.MergeStage;
 import io.mohamed.resolver.core.model.Dependency;
